#7788 - Editor revizí - oprávnění - schválené entity může měnit uživatel bez oprávnění Změna schválených archivních entit

+++ elza/elza-core/src/main/java/cz/tacr/elza/controller/AccessPointController.java
@@ -6,12 +6,6 @@ import java.util.List;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 
-import cz.tacr.elza.controller.vo.RevStateChange;
-import cz.tacr.elza.core.data.ItemType;
-import cz.tacr.elza.core.data.StaticDataProvider;
-import cz.tacr.elza.core.data.StaticDataService;
-import cz.tacr.elza.service.RevisionService;
-
 import org.apache.commons.lang3.Validate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +16,10 @@ import cz.tacr.elza.controller.vo.AutoValue;
 import cz.tacr.elza.controller.vo.DeleteAccessPointDetail;
 import cz.tacr.elza.controller.vo.DeleteAccessPointsDetail;
 import cz.tacr.elza.controller.vo.ResultAutoItems;
+import cz.tacr.elza.controller.vo.RevStateChange;
+import cz.tacr.elza.core.data.ItemType;
+import cz.tacr.elza.core.data.StaticDataProvider;
+import cz.tacr.elza.core.data.StaticDataService;
 import cz.tacr.elza.domain.ApAccessPoint;
 import cz.tacr.elza.domain.ApRevision;
 import cz.tacr.elza.domain.ApState;
@@ -32,6 +30,7 @@ import cz.tacr.elza.exception.codes.RegistryCode;
 import cz.tacr.elza.groovy.GroovyItem;
 import cz.tacr.elza.service.AccessPointService;
 import cz.tacr.elza.service.GroovyService;
+import cz.tacr.elza.service.RevisionService;
 
 @RestController
 @RequestMapping("/api/v1")
@@ -81,6 +80,7 @@ public class AccessPointController implements AccesspointsApi {
     @Transactional
     public ResponseEntity<Void> createRevision(Integer id) {
         ApState state = accessPointService.getStateInternal(id);
+        accessPointService.checkPermissionForEditingConfirmed(state);
 
         // Nelze vytvořit revizi, pokud má archivní entita jiný stav než NEW, TO_AMEND nebo APPROVED
         if (!Arrays.asList(StateApproval.NEW, StateApproval.TO_AMEND, StateApproval.APPROVED).contains(state.getStateApproval())) {
@@ -96,6 +96,8 @@ public class AccessPointController implements AccesspointsApi {
     @Transactional
     public ResponseEntity<Void> deleteRevision(Integer id) {
         ApState state = accessPointService.getStateInternal(id);
+        accessPointService.checkPermissionForEditingConfirmed(state);
+
         revisionService.deleteRevision(state);
         return ResponseEntity.ok().build();
     }
@@ -112,6 +114,8 @@ public class AccessPointController implements AccesspointsApi {
     @Transactional
     public ResponseEntity<Void> deleteRevisionPart(Integer id, Integer partId) {
         ApState state = accessPointService.getStateInternal(id);
+        accessPointService.checkPermissionForEditingConfirmed(state);
+
         revisionService.deletePart(state, partId);
         return ResponseEntity.ok().build();
     }
